requirements.txt
asgiref==3.6.0
Beaker==1.12.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.6
django-cors-headers==3.7.0
django-forest==0.1.1
django-ipware==3.0.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
docopt==0.6.2
ecdsa==0.18.0
future==0.18.3
idna==2.10
importlib-metadata==4.6.3
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-jsonapi==0.24.0
oic==1.2.1
packaging==23.0
Pillow==9.4.0
pipreqs==0.4.11
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.17
pyjwkest==1.4.2
PyJWT==2.6.0
python-dateutil==2.8.1
python-jose==3.3.0
pytz==2022.7.1
requests==2.25.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
yarg==0.1.9
zipp==3.12.1
